refactor(webapp): migrate label/Tag to new v2 Tag component
Create components-v2/ui/Tag.tsx with the same semantic variant names
(success, alert, info, warning, gray, gray1, neutral) but backed by v2
design tokens (feedback-* and badge-* CSS vars). Update all 7 callsites
to import from the new location and delete the v1 source file.
