+++ packages/webapp/src/pages/Logs/Operation/constants.tsx
@@ -1,6 +1,6 @@
 import { IconChevronRight } from '@tabler/icons-react';
 
-import { Tag } from '../../../components/ui/label/Tag';
+import { Tag } from '@/components-v2/ui/Tag';
 import { formatDateToLogFormat, millisecondsToRuntime } from '../../../utils/utils';
 import { LevelTag } from '../components/LevelTag';
